adding rec quick_sort

+++ 2-selection_sort.c
@@ -10,7 +10,7 @@
 void selection_sort(int *array, size_t size)
 {
 	size_t i, j;
-	int temp;
+	size_t temp;
 	size_t smaller_elem;
 
 	for (i = 0; i < size - 1; i++)
